tests/requirements.txt
azure-cli-core==2.37.0
azure-cli-telemetry==1.0.6
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-mgmt-datafactory==2.10.0
azure-mgmt-resource==21.2.1
azure-mgmt-synapse==2.1.0b7
azure-storage-blob==12.9.0
pytest==6.2.5
pytest-cov==2.11.1
pyodbc